+++ src/ezdxf/entities/__init__.py
@@ -99,6 +99,7 @@ from .mleader import MLeader, MLeaderStyle, MLeaderStyleCollection, MultiLeader
 # register graphical entities R2007
 
 from .light import Light
+from .acad_table import AcadTableBlockContent
 
 # register graphical entities R2010