+++ NEWS.md
@@ -6,6 +6,8 @@ Version 0.17.2b1 - dev
 ----------------------
 
 - BUGFIX: flattening issues #574 in Path() and ConstructionEllipse() 
+- BUGFIX: `drawing` add-on shows block references in `ACAD_TABLE` at the 
+  correct location  
 - PREVIEW: arc length dimension rendering support, new factory methods: 
   `add_arc_dim_3p()`, `add_arc_dim_cra()`, `add_arc_dim_arc()`
 - PREVIEW: ordinate dimension rendering support, new factory methods: